requirements.txt
google-adk>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
huggingface_hub>=0.14.0

